# src/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Ecomm",
    "contact": {}
  },
  "host": "localhost:7600",
  "basePath": "/api",
  "securityDefinitions": {},
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "signUp",
        "tags": [
          "Auth"
        ],
        "operationId": "signUp",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/user/signIn": {
      "post": {
        "summary": "signIn",
        "tags": [
          "Auth"
        ],
        "operationId": "signIn",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": false,
            "default": "Bearer {token}",
            "type": "string"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/signInRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        }
      }
    },
    "/products/addProduct": {
      "post": {
        "summary": "addProduct",
        "tags": [
          "Products"
        ],
        "operationId": "addProduct",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/addProductRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/products/getAllProduct": {
      "get": {
        "summary": "getAllProduct",
        "tags": [
          "Products"
        ],
        "operationId": "getAllProduct",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/products/deleteProduct/62be98f23998929a187abc1c": {
      "post": {
        "summary": "deleteProduct",
        "tags": [
          "Products"
        ],
        "operationId": "deleteProduct",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/products/updateProduct/62be9762e4db098a905dcc2d": {
      "post": {
        "summary": "updateProduct",
        "tags": [
          "Products"
        ],
        "operationId": "updateProduct",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updateProductRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/cart/addToCart": {
      "post": {
        "summary": "addToCart",
        "tags": [
          "cart"
        ],
        "operationId": "addToCart",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/addToCartRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/cart/getCartItem": {
      "get": {
        "summary": "getCartItem",
        "tags": [
          "cart"
        ],
        "operationId": "getCartItem",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/cart/removeCartItem/62beb8f5439cff38bdd568df": {
      "post": {
        "summary": "removeCartItem",
        "tags": [
          "cart"
        ],
        "operationId": "removeCartItem",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/invoice/createInvoice/62beefe61af610d4c26b1624": {
      "post": {
        "summary": "createInvoice",
        "tags": [
          "Invoice"
        ],
        "operationId": "createInvoice",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/createInvoiceRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/invoice/getAllInvoices": {
      "get": {
        "summary": "getAllInvoices",
        "tags": [
          "Invoice"
        ],
        "operationId": "getAllInvoices",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/invoice/deleteInvoice/62c277e6a88706cb5f1d9928": {
      "post": {
        "summary": "deleteInvoice",
        "tags": [
          "Invoice"
        ],
        "operationId": "deleteInvoice",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/invoice/updateInvoice/62c277e6a88706cb5f1d9928": {
      "post": {
        "summary": "updateInvoice",
        "tags": [
          "Invoice"
        ],
        "operationId": "updateInvoice",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/updateInvoiceRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "security": []
      }
    }
  },
  "definitions": {
    "signInRequest": {
      "title": "signInRequest",
      "example": {
        "email": "ketan@7600",
        "password": "7600"
      },
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "addProductRequest": {
      "title": "addProductRequest",
      "example": {
        "title": "Laptop",
        "price": 35000,
        "description": "Good Porduct",
        "quantity": 50
      },
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "price": {
          "type": "integer",
          "format": "int32"
        },
        "description": {
          "type": "string"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "title",
        "price",
        "description",
        "quantity"
      ]
    },
    "updateProductRequest": {
      "title": "updateProductRequest",
      "example": {
        "title": "Mi",
        "price": 50000,
        "description": "Good Porduct",
        "quantity": 50
      },
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "price": {
          "type": "integer",
          "format": "int32"
        },
        "description": {
          "type": "string"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "title",
        "price",
        "description",
        "quantity"
      ]
    },
    "addToCartRequest": {
      "title": "addToCartRequest",
      "example": {
        "item": "62c2869c74a6a4a5cb6b41ae",
        "quantity": 1
      },
      "type": "object",
      "properties": {
        "item": {
          "type": "string"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        }
      },
      "required": [
        "item",
        "quantity"
      ]
    },
    "createInvoiceRequest": {
      "title": "createInvoiceRequest",
      "example": {
        "contact": 1236547895,
        "address": "Udhana",
        "city": "Surat",
        "state": "Gujarat"
      },
      "type": "object",
      "properties": {
        "contact": {
          "type": "integer",
          "format": "int32"
        },
        "address": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "state": {
          "type": "string"
        }
      },
      "required": [
        "contact",
        "address",
        "city",
        "state"
      ]
    },
    "updateInvoiceRequest": {
      "title": "updateInvoiceRequest",
      "example": {
        "contact": 8140835271,
        "address": "Udhana",
        "city": "Surat",
        "state": "Gujarat"
      },
      "type": "object",
      "properties": {
        "contact": {
          "type": "integer",
          "format": "int64"
        },
        "address": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "state": {
          "type": "string"
        }
      },
      "required": [
        "contact",
        "address",
        "city",
        "state"
      ]
    }
  },
  "security": [],
  "tags": [
    {
      "name": "Auth"
    },
    {
      "name": "Products"
    },
    {
      "name": "cart"
    },
    {
      "name": "Invoice"
    }
  ]
}